requirements.txt
sentence-transformers==3.0.1
transformers==4.41.0
torch==2.3.0 --index-url https://download.pytorch.org/whl/cpu
trafilatura==1.6.3
beautifulsoup4==4.12.2
lxml_html_clean==0.1.0
requests==2.31.0
tavily-python==0.3.0
pydantic==2.5.3
python-dotenv==1.0.0
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
keybert==0.8.3
nltk==3.8.1
regex==2023.10.3
aiohttp==3.9.1
tqdm==4.66.1
duckduckgo-search>=5.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pytest==7.4.3
flask-limiter==3.5.0
whitenoise==6.6.0
